refactor: rehome MAIN runtime infrastructure
