+++ src/content/main/infrastructure/history-diagnostics.js
