+++ src/content/main/infrastructure/operation-guard.js
