+++ src/features/batch-section-creation-history-record.js
@@ -1,5 +1,5 @@
 import * as modelApi from './batch-section-creation-history-model.js';
-import { historyDiagnostics } from '../shared/history-diagnostics.js';
+import { historyDiagnostics } from '../content/main/infrastructure/history-diagnostics.js';
 
 const TERMINAL_STATUSES = new Set(modelApi.TERMINAL_STATUSES);
