+++ src/features/batch-user-management-history.js
@@ -1,4 +1,7 @@
-import { diagnosticsFromAttempts, historyDiagnostics } from '../shared/history-diagnostics.js';
+import {
+    diagnosticsFromAttempts,
+    historyDiagnostics
+} from '../content/main/infrastructure/history-diagnostics.js';
 
 const OPERATION_TYPE = 'batch_user_management';
 const OPERATION_NAMES = Object.freeze({
